+++ R/itemfit.R
@@ -418,8 +418,13 @@ itemfit <- function(x, Zh = TRUE, X2 = FALSE, group.size = 150, mincell = 1, S_X
         S_X2 <- df.S_X2 <- numeric(J)
         O <- makeObstables(dat, x@K)
         Nk <- rowSums(O[[1L]])
+        dots <- list(...)
+        quadpts <- dots$quadpts
+        theta_lim <- dots$theta_lim
+        if(is.null(quadpts)) quadpts <- 61
+        if(is.null(theta_lim)) theta_lim <- c(-6,6)
         E <- EAPsum(x, S_X2 = TRUE, gp = list(gmeans=0, gcov=matrix(1)), CUSTOM.IND=x@CUSTOM.IND,
-                    ...)
+                    quadpts=quadpts, theta_lim=theta_lim)
         for(i in 1L:J)
             E[[i]] <- E[[i]] * Nk
         coll <- collapseCells(O, E, mincell=mincell)